Move successful period creation tests from serializer tests to views tests
