+++ budget_api/periods/tests/test_api.py
@@ -1,0 +1,129 @@
+from datetime import date
+from typing import Any
+
+import pytest
+from django.urls import reverse
+from factory.base import FactoryMetaClass
+from periods.models import BudgetingPeriod
+from periods.serializers import BudgetingPeriodSerializer
+from rest_framework import status
+from rest_framework.test import APIClient
+
+PERIODS_URL = reverse('periods:budgetingperiod-list')
+
+
+def period_detail_url(period_id):
+    """Create and return a budgeting period detail URL."""
+    return reverse('recipe:budgetingperiod-detail', args=[period_id])
+
+
+def check_period_validated_data(response, payload):
+    """Helper function for checking BudgetingPeriod validated data."""
+    instance = BudgetingPeriod.objects.get(id=response.data['id'])
+    for key in payload:
+        if key == 'user':
+            assert getattr(getattr(instance, key), 'id') == payload[key]
+        else:
+            assert getattr(instance, key) == payload[key]
+
+
+@pytest.mark.django_db
+class TestBudgetingPeriodViewSet:
+    """Tests for BudgetingPeriodViewSet."""
+
+    def test_auth_required(self, api_client: APIClient):
+        """Test auth is required to call endpoint."""
+        res = api_client.get(PERIODS_URL)
+
+        assert res.status_code == status.HTTP_401_UNAUTHORIZED
+
+    def test_retrieve_periods_list(
+        self, api_client: APIClient, base_user: Any, budgeting_period_factory: FactoryMetaClass
+    ):
+        """Test retrieving list of periods."""
+        api_client.force_authenticate(base_user)
+        budgeting_period_factory(
+            user=base_user, date_start=date(2023, 1, 1), date_end=date(2023, 1, 31), is_active=False
+        )
+        budgeting_period_factory(
+            user=base_user, date_start=date(2023, 2, 1), date_end=date(2023, 2, 28), is_active=True
+        )
+
+        response = api_client.get(PERIODS_URL)
+
+        periods = BudgetingPeriod.objects.all().order_by('-date_start')
+        serializer = BudgetingPeriodSerializer(periods, many=True)
+        assert response.status_code == status.HTTP_200_OK
+        assert response.data['results'] == serializer.data
+
+    def test_periods_list_limited_to_user(
+        self, api_client: APIClient, user_factory: FactoryMetaClass, budgeting_period_factory: FactoryMetaClass
+    ):
+        """Test retrieved list of periods is limited to authenticated user."""
+        user = user_factory()
+        budgeting_period_factory(user=user)
+        budgeting_period_factory()
+        api_client.force_authenticate(user)
+
+        response = api_client.get(PERIODS_URL)
+
+        periods = BudgetingPeriod.objects.filter(user=user)
+        serializer = BudgetingPeriodSerializer(periods, many=True)
+        assert response.status_code == status.HTTP_200_OK
+        assert response.data['results'] == serializer.data
+
+    def test_create_single_period(self, api_client: APIClient, base_user: Any):
+        """Test creating single BudgetingPeriod."""
+        api_client.force_authenticate(base_user)
+        payload = {'name': '2023_01', 'date_start': date(2023, 1, 1), 'date_end': date(2023, 1, 31)}
+
+        response = api_client.post(PERIODS_URL, payload)
+
+        assert response.status_code == status.HTTP_201_CREATED
+        assert BudgetingPeriod.objects.filter(user=base_user).count() == 1
+        check_period_validated_data(response, payload)
+        period = BudgetingPeriod.objects.get(id=response.data['id'])
+        serializer = BudgetingPeriodSerializer(period)
+        assert response.data == serializer.data
+
+    def test_create_two_periods_by_one_user(self, api_client: APIClient, base_user: Any):
+        """Test creating two valid BudgetingPeriods by single user."""
+        api_client.force_authenticate(base_user)
+        payload_1 = {
+            'name': '2023_01',
+            'date_start': date(2023, 1, 1),
+            'date_end': date(2023, 1, 31),
+        }
+        payload_2 = {
+            'name': '2023_02',
+            'date_start': date(2023, 2, 1),
+            'date_end': date(2023, 2, 28),
+        }
+
+        response_1 = api_client.post(PERIODS_URL, payload_1)
+        response_2 = api_client.post(PERIODS_URL, payload_2)
+
+        assert response_1.status_code == status.HTTP_201_CREATED
+        assert response_2.status_code == status.HTTP_201_CREATED
+        assert BudgetingPeriod.objects.filter(user=base_user).count() == 2
+        for response, payload in [(response_1, payload_1), (response_2, payload_2)]:
+            check_period_validated_data(response, payload)
+
+    def test_create_same_period_by_two_users(self, api_client: APIClient, user_factory):
+        """Test creating period with the same params by two users."""
+        payload = {
+            'name': '2023_01',
+            'date_start': date(2023, 1, 1),
+            'date_end': date(2023, 1, 31),
+        }
+        user_1 = user_factory()
+        api_client.force_authenticate(user_1)
+        api_client.post(PERIODS_URL, payload)
+
+        user_2 = user_factory()
+        api_client.force_authenticate(user_2)
+        api_client.post(PERIODS_URL, payload)
+
+        assert BudgetingPeriod.objects.all().count() == 2
+        assert BudgetingPeriod.objects.filter(user=user_1).count() == 1
+        assert BudgetingPeriod.objects.filter(user=user_2).count() == 1
